pyproject.toml
[project]
name = "ambiente-virtual-discord-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Rukasuleonhart",email = "lucasoliveirasantossoares@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
